requirements.txt
PyQt6>=6.6.0
httpx>=0.27.0
pyinstaller>=6.0.0
flask>=3.0.0
flask-cors>=4.0.0
pywebview>=5.0.0
pystray>=0.19.0
Pillow>=10.0.0
